refactor: move PopularMovieResponse to its own package
