+++ data/src/test/java/com/london/data/mapper/popular/PopularMovieMapperTest.kt
@@ -2,7 +2,7 @@ package com.london.data.mapper.popular
 
 import com.google.common.truth.Truth.assertThat
 import com.london.data.remote.model.ApiResponse
-import com.london.data.remote.model.home.model.PopularMovieResponse
+import com.london.data.remote.model.home.model.popular.PopularMovieResponse
 import com.london.domain.entity.popular.PopularMovie
 import org.junit.Test
